requirements.txt
tensorflow-gpu
keras
graphviz
pydot
tqdm
opencv-python
numpy